// sportpartner-spa-web/src/app/modules/chat/components/message-detail-item/message-type.enum.ts
export enum MessageTypeEnum {
    FirstMessage = 'FirstMessage',
    HighFive = 'HighFive',
    HighFiveBack = 'HighFiveBack',
    Normal = 'Normal',
    sayHi = 'SayHi',
    helloEveryone = 'HelloEveryone',
    joinConversation = 'JoinConversation',
    leftConversation = 'LeftConversation',
    welcomeGroup = 'WelcomeGroup',
}