// sportpartner-spa-web/src/app/const.enum.ts
export const enum DeviceTypes {
    desktop = 'desktop',
    tablet = 'tablet',
    mobile = 'mobile'
}

export enum PanelType {
    USER_PROFILE = 'PANEL_USER_PROFILE',
    GROUP_INFO = 'PANEL_GROUP_INFO',
}

export enum MessageStatus {
    SENDING = 'SENDING',
    SENT = 'SENT',
    FAIL = 'FAIL'
}